backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.1
Werkzeug==3.1.3
